test: move test.
