+++ tests/test_flatland_envs_rail_grid_transition_map.py
@@ -1,6 +1,8 @@
+import numpy as np
 import pytest
 
 from flatland.core.grid.grid4 import Grid4TransitionsEnum
+from flatland.core.grid.grid4_utils import get_new_position
 from flatland.envs.grid.rail_env_grid import RailEnvTransitionsEnum, RailEnvTransitions
 from flatland.envs.rail_env_action import RailEnvActions
 from flatland.envs.rail_grid_transition_map import RailGridTransitionMap
@@ -596,3 +598,64 @@ def test_is_valid_configuration_out_of_bounds(configuration, expected):
     rail = RailGridTransitionMap(3, 3, RailEnvTransitions())
     rail.set_transitions((0, 0), RailEnvTransitionsEnum.horizontal_straight)
     assert rail.is_valid_configuration(configuration) == expected
+
+
+_NON_EMPTY_RAIL_ENV_TRANSITIONS = [t for t in RailEnvTransitionsEnum if t != RailEnvTransitionsEnum.empty]
+
+
+@pytest.mark.parametrize("rail_env_transition", _NON_EMPTY_RAIL_ENV_TRANSITIONS,
+                         ids=[t.name for t in _NON_EMPTY_RAIL_ENV_TRANSITIONS])
+@pytest.mark.parametrize("direction", list(Grid4TransitionsEnum), ids=[d.name for d in Grid4TransitionsEnum])
+def test_apply_action_independent_not_none_for_every_entry_side(rail_env_transition, direction):
+    """For every non-empty RailEnvTransitionsEnum element placed at the center of a 3x3 grid, surrounded by a
+    straight track on every side from which the element itself has an outgoing transition (i.e. a side that
+    get_predecessor_configurations reports as a valid entry), apply_action_independent() finds at least one
+    valid action (MOVE_LEFT/MOVE_FORWARD/MOVE_RIGHT) for the configuration reached by entering from that side."""
+    transitions = RailEnvTransitions()
+    center = (1, 1)
+    grid = np.zeros((3, 3), dtype=np.uint16)
+    grid[center] = rail_env_transition
+    for d in Grid4TransitionsEnum:
+        if not any(transitions.get_transitions(int(rail_env_transition), d)):
+            continue  # element has no transition for this incoming orientation -- no neighbor to connect
+        previous_cell = get_new_position(center, (d + 2) % 4)
+        is_vertical = d in (Grid4TransitionsEnum.NORTH, Grid4TransitionsEnum.SOUTH)
+        sender = RailEnvTransitionsEnum.vertical_straight if is_vertical else RailEnvTransitionsEnum.horizontal_straight
+        grid[previous_cell] = sender
+
+    rail = RailGridTransitionMap(width=3, height=3, transitions=transitions)
+    rail.grid = grid
+
+    configuration = (center, direction)
+    if not rail.get_predecessor_configurations(configuration):
+        pytest.skip(f"{rail_env_transition.name} has no valid entry from {direction.name}")
+
+    assert any(rail.apply_action_independent(action, configuration) is not None
+               for action in (RailEnvActions.MOVE_LEFT, RailEnvActions.MOVE_FORWARD, RailEnvActions.MOVE_RIGHT))
+
+
+@pytest.mark.parametrize("action", list(RailEnvActions), ids=[a.name for a in RailEnvActions])
+def test_apply_action_independent_only_left_right_valid_at_symmetric_switch(action):
+    """Same grid as test_env_collision_penalty_on_invalid_forward_at_symmetric_switch: at the switch (2,1),
+    heading WEST, only north/south transitions exist, so only MOVE_LEFT/MOVE_RIGHT are valid actions --
+    apply_action_independent() returns None for every other action (DO_NOTHING, MOVE_FORWARD, STOP_MOVING)."""
+    transitions = RailEnvTransitions()
+    grid = np.zeros((5, 6), dtype=np.uint16)
+    grid[2, 1] = RailEnvTransitionsEnum.symmetric_switch_from_east  # heading west forks N/S
+    grid[2, 2] = RailEnvTransitionsEnum.horizontal_straight
+    grid[2, 3] = RailEnvTransitionsEnum.horizontal_straight
+    grid[2, 4] = RailEnvTransitionsEnum.dead_end_from_east
+    grid[1, 1] = RailEnvTransitionsEnum.vertical_straight
+    grid[0, 1] = RailEnvTransitionsEnum.dead_end_from_north
+    grid[3, 1] = RailEnvTransitionsEnum.vertical_straight
+    grid[4, 1] = RailEnvTransitionsEnum.dead_end_from_south
+
+    rail = RailGridTransitionMap(width=6, height=5, transitions=transitions)
+    rail.grid = grid
+
+    configuration = ((2, 1), Grid4TransitionsEnum.WEST)
+    result = rail.apply_action_independent(action, configuration)
+    if action in (RailEnvActions.MOVE_LEFT, RailEnvActions.MOVE_RIGHT):
+        assert result is not None
+    else:
+        assert result is None
